Squashed commit of the following:
commit d5b11f60d5f385a4bf7921cd8f0410945af248f2
Author: Reis <80530563+JoaoReis153@users.noreply.github.com>
Date:   Sat Apr 18 19:29:06 2026 +0100

    Bruno tests remake (#91)

    * TEST: add Bruno REST smoke collection and make target

    * CI: add GitHub Actions workflow for Bruno REST tests

    * AUTH: map duplicate user to conflict response

    * CHORE: ignore generated Bruno reports

    * ci: align bruno workflow with make target

    * ci: test to see if actions run

    * ci: added run of bruno tests on pushes

    * ci: testing actions

    * ci: harden bruno workflow artifact handling

    * ci: add actions debug for gateway readiness

    * ci: fix auth jwt secrets in bruno workflow

    * del: unused postman references

    * refactor: notifications bruno tests

    * refactor: restaurants bruno tests

    * refactor: reservations bruno tests

    * refactor: users bruno tests

    * refactor: sponsor bruno tests

    * refactor: reviews bruno tests

    * chore: removal of unsafe log

    * chore: added variable to bruno tests

    * fix: removed malformed test

    * fix: added missing env to bruno test

commit de6e2a6a40b5f23844de0ce1eed4cc9446f399c9
Author: Reis <80530563+JoaoReis153@users.noreply.github.com>
Date:   Sat Apr 18 17:21:00 2026 +0100

    Data loading and setup (#79)

    * FEAT: add CSV generation pipeline for auth restaurant booking

    * CHORE: add generated processed CSV seed outputs

    * DOCS: add setup data generation and loading commands

    * CHORE: add AUTH_POSTGRES_DB to env template

    * FEAT: update booking and restaurant schema constraints

    * CHORE: update ignore rules for data and cache

    * REFACTOR: split service-specific CSV logic and improve streaming paths

    * FIX: resolve generator mismatch in CSV pipeline

    * PERF: speed up reviews loading and category indexing

    * PERF: stream restaurant generation directly from CSV files without DataFrames

    * DEL: processed data

    * CHORE: renamed the files to bug fix

    * DATA: Use source gPlus IDs for users and restaurants

    * UPD: Ignore zip artifacts

    * CHORE: add create_env workflow and simplify env template

    * DATA: Replace booking CSV generation with review pipeline and align seed auth password hash

    * DATA: Add load-reviews target and switch seed IDs to DB-safe integers

    * CHORE: update the makefile to load data into container

    * DATA: Expand test coverage to all microservices and verify system validation

    * DATA: Align CSV IDs with DB schema across services

    * CHORE: Simplify README setup instructions

    * FIX: README.md

    * FIX: Generate user_id on auth register insert

    * FIX: Generate restaurant_id on restaurant register insert

    * DATA: Hash gPlus place and user IDs to BIGINT in scripts

    * DATA: Migrate service schemas from NUMERIC IDs to BIGINT

    * DATA: Document hash-based ID strategy

    * DOCS: Document scripts process structure

    * DOCS: Add ID strategy heading

    * FIX: Migrate user and restaurant IDs to int64 across services

    * fix: updated owner id to bigInt

    * chore: added '-v' tag to down command in Makefile

    * test: added and updated collections for restarant service

    * Fix: Change review_id from int32 to int64

    * fix: restaurant details connection to reviews

    * chore: added more make commands

    * chore: removed exposes in docker compose files of all services

    ---------

+++ services/review/db_setup.sql
@@ -3,8 +3,8 @@ DROP TABLE IF EXISTS restaurant_stats;
 
 CREATE TABLE IF NOT EXISTS review(
     review_id     SERIAL PRIMARY KEY,
-    restaurant_id INT NOT NULL,
-    user_id       INT NOT NULL,
+    restaurant_id BIGINT NOT NULL,
+    user_id       BIGINT NOT NULL,
     comment       VARCHAR(100) NOT NULL,
     rating        INT NOT NULL CHECK (rating >= 1 AND rating <= 5),
     created_at    TIMESTAMP DEFAULT NOW() NOT NULL,
@@ -12,7 +12,7 @@ CREATE TABLE IF NOT EXISTS review(
 );
 
 CREATE TABLE IF NOT EXISTS restaurant_stats(
-    restaurant_id  INT PRIMARY KEY,
+    restaurant_id  BIGINT PRIMARY KEY,
     average_rating DECIMAL(3,2) NOT NULL,
     review_count INT NOT NULL
 );
@@ -21,7 +21,7 @@ CREATE INDEX IF NOT EXISTS idx_review_restaurant_rating ON review(restaurant_id,
 
 CREATE OR REPLACE FUNCTION update_restaurant_stats() RETURNS TRIGGER AS $$
 DECLARE
-    target_id INT;
+    target_id BIGINT;
     v_avg DECIMAL(3,2);
     v_count INT;
 BEGIN
